+++ resources/migrations/20210810150445-invitations.up.sql
@@ -1,0 +1,4 @@
+CREATE TABLE invitation (
+  id serial NOT NULL PRIMARY KEY,
+  invitationdata jsonb NOT NULL
+)